feat: gateway log display for admin

+++ src/Controllers/Admin/UserController.php
@@ -171,9 +171,9 @@ final class UserController extends BaseController
 
         if ($request->getParam('money') !== '' &&
             $request->getParam('money') !== null &&
-            (double) $request->getParam('money') !== (float) $user->money
+            (float) $request->getParam('money') !== (float) $user->money
         ) {
-            $money = (double) $request->getParam('money');
+            $money = (float) $request->getParam('money');
             $diff = $money - $user->money;
             $remark = ($diff > 0 ? '管理员添加余额' : '管理员扣除余额');
             (new UserMoneyLog())->addMoneyLog($id, (float) $user->money, $money, $diff, $remark);
